REFACTOR: criando pasta crypto-engine para organizar as classes e funções do antigo crypto.ts #4
